[REFACTOR] Background Color 코드 맨 앞에 # 추가 - #1
