+++ admin-api/src/main/kotlin/com/woomulwoomul/adminapi/service/question/response/QuestionFindResponse.kt
@@ -1,6 +1,6 @@
 package com.woomulwoomul.adminapi.service.question.response
 
-import com.woomulwoomul.core.common.constant.BackgroundColor
+import com.woomulwoomul.core.domain.question.BackgroundColor
 import com.woomulwoomul.core.domain.base.ServiceStatus
 import com.woomulwoomul.core.domain.question.CategoryEntity
 import com.woomulwoomul.core.domain.question.QuestionEntity
