+++ admin-api/src/main/resources/templates/question/question-new.html
@@ -43,7 +43,7 @@
                            th:value="${availableBackgroundColor.getValue()}"
                            class="form-check-input" />
                     <span th:text="${availableBackgroundColor.getValue()}"
-                          th:style="'background-color: #' + ${availableBackgroundColor.getValue()}"
+                          th:style="'background-color: ' + ${availableBackgroundColor.getValue()}"
                           class="form-check-label"></span>
                 </label>
             </td>
