+++ admin-api/src/main/resources/templates/question/question.html
@@ -56,7 +56,7 @@
                            th:checked="${response.backgroundColor.equals(availableBackgroundColor)}"
                            class="form-check-input" />
                     <span th:text="${availableBackgroundColor.getValue()}"
-                          th:style="'background-color: #' + ${availableBackgroundColor.getValue()}"
+                          th:style="'background-color: ' + ${availableBackgroundColor.getValue()}"
                     class="form-check-label"></span>
                 </label>
             </td>
