+++ admin-api/src/main/resources/templates/question/questions.html
@@ -53,7 +53,7 @@
                 <td th:text="${question.text}">내용</td>
                 <td>
                     <span th:text="${question.backgroundColor.getValue()}"
-                          th:style="'background-color: #' + ${question.backgroundColor.getValue()}"
+                          th:style="'background-color: ' + ${question.backgroundColor.getValue()}"
                           class="form-check-label"></span>
                 </td>
                 <td th:text="${#temporals.format(question.startDateTime, 'yyyy-MM-dd HH:mm:ss')}">시작일</td>
