+++ client-api/src/docs/asciidoc/api/question/create-user-question.adoc
@@ -33,7 +33,7 @@ include::{snippets}/question/create-user-question/response-fields.adoc[]
 ----
 {
   "code" : "QUESTION_BACKGROUND_COLOR_PATTERN_INVALID",
-  "message" : "질문 배경 색상은 'FFACA8', 'FFA34F', 'FFC34F', 'C5FFAA', '1AE7D8', '4FB5FF', '868BFF', 'C58AFF', 'FF9CE3', 'FFFFFF' 중 하나만 가능합니다."
+  "message" : "질문 배경 색상은 '#FFACA8', '#FFA34F', '#FFC34F', '#C5FFAA', '#1AE7D8', '#4FB5FF', '#868BFF', '#C58AFF', '#FF9CE3', '#FFFFFF' 중 하나만 가능합니다."
 }
 ----
 ----
