+++ client-api/src/main/kotlin/com/woomulwoomul/clientapi/service/question/response/AnswerCreateResponse.kt
@@ -1,6 +1,5 @@
 package com.woomulwoomul.clientapi.service.question.response
 
-import com.woomulwoomul.core.common.constant.BackgroundColor
 import com.woomulwoomul.core.domain.question.CategoryEntity
 import com.woomulwoomul.core.domain.question.QuestionEntity
 import com.woomulwoomul.core.domain.user.UserEntity
