+++ client-api/src/test/kotlin/com/woomulwoomul/clientapi/service/question/AnswerServiceTest.kt
@@ -3,7 +3,7 @@ package com.woomulwoomul.clientapi.service.question
 import com.woomulwoomul.clientapi.service.question.request.AnswerCreateServiceRequest
 import com.woomulwoomul.clientapi.service.question.request.AnswerUpdateServiceRequest
 import com.woomulwoomul.clientapi.service.question.response.AnswerFindAllCategoryResponse
-import com.woomulwoomul.core.common.constant.BackgroundColor
+import com.woomulwoomul.core.domain.question.BackgroundColor
 import com.woomulwoomul.core.common.constant.ExceptionCode.*
 import com.woomulwoomul.core.common.request.PageCursorRequest
 import com.woomulwoomul.core.common.response.CustomException
