+++ core/src/test/kotlin/com/woomulwoomul/core/domain/question/QuestionCategoryEntityTest.kt
@@ -1,6 +1,5 @@
 package com.woomulwoomul.core.domain.question
 
-import com.woomulwoomul.core.common.constant.BackgroundColor
 import com.woomulwoomul.core.domain.base.ServiceStatus
 import com.woomulwoomul.core.domain.user.UserEntity
 import org.assertj.core.api.Assertions.assertThat
